feat: add some magic to errors
