+++ src/app/(dashboard)/test-page/components/TestErrorPage/index.ts
@@ -1,0 +1,1 @@
+export { TestErrorPage } from "./TestErrorPage";
