+++ src/app/(dashboard)/test-page/page.tsx
@@ -1,0 +1,12 @@
+import type { Metadata } from "next";
+import { TestErrorPage } from "./components/TestErrorPage";
+
+export async function generateMetadata(): Promise<Metadata> {
+  return {
+    title: "Test Page - Tipi",
+  };
+}
+
+export default function TestPage() {
+  return <TestErrorPage />;
+}
