+++ src/client/messages/en.json
@@ -224,6 +224,7 @@
   "HEADER_SOURCE_CODE": "Source code",
   "HEADER_SPONSOR": "Sponsor",
   "HEADER_UPDATE_AVAILABLE": "Update available",
+  "HEADER_TEST_PAGE": "Test Page",
   "INTERNAL_SERVER_ERROR": "Internal server error",
   "LINKS_ADD_SUBMIT": "Submit",
   "LINKS_ADD_SUBTITLE": "Add external link to the dashboard",
